// src/api/__generated__/story/model/dramaPlayResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Service API
 * Mini Drama Platform
 * OpenAPI spec version: v1.0
 */

export interface DramaPlayResponse {
  dramaId?: number;
  episodeId?: number;
  episodeNo?: number;
  title?: string;
  coverUrl?: string;
  description?: string;
  mediaAccessUrl?: string;
  playbackType?: string;
  likeCount?: number;
  commentCount?: number;
  favoriteCount?: number;
  favoritedByMe?: boolean;
  likedByMe?: boolean;
  videoUrl?: string;
  creatorId?: number;
  creatorName?: string;
  creatorAvatarUrl?: string;
}